Update preprocessing.zsh
added (if-if) safegaurds for trimming and alignment failure so progress is not deleted with rm -rf

+++ preprocessing.zsh
@@ -4,24 +4,25 @@
 # Creator: Solana Fernandez
 ##################################################################################################
 ##################################################################################################
-################    NOTE: export $GET_DATA and $DIRECTORYbeforehand 	  ########################
-################	             				          ########################
+################    NOTE: export $GET_DATA and $DIRECTORY beforehand 	  ########################
+################	    PLEASE EXECUTE IN CONDA ENVIRONMENT	          ########################
 ################                                                          ########################
 ################  EXAMPLE USAGE (FEEL FREE TO UNCOMMENT AND RUN with):    ########################
 ################       (no spaces between '=' when defining)    	  ########################
-# export DIRECTORY="/Users/solanafernandez/Desktop/UCSD_local/BENG204/project" # desired location of file dump
-# export GET_DATA="zsh/test1.sh" # 	location of wget bash script
+
+# export DIRECTORY="/Users/solanafernandez/Desktop/UCSD_local/BENG204/project"  # desired location of file dump
+# export GET_DATA="zsh/test1.sh" 						# location of wget bash script
 
 ##################################################################################################
 ######## --------------> $DIRECTORY should be the path that contains zsh data retrieval folder
 cd $DIRECTORY
 
-##########################  Retrieve Data: Test (make sure executable) ############################
+##########################  Retrieve Data (make sure executable) #################################
 chmod +x $GET_DATA
 
 zsh $GET_DATA # throws fastq's in current directory (deleted later)
 
-##########################  FastQC (input: fastq; output: fastq) ##################################  
+######################################  FastQC  ###################################################  
 mkdir fastqc
 find . -name "*.fastq.gz" | xargs fastqc -t 8 -o fastqc
 
@@ -38,8 +39,11 @@ sort 2_end.txt -o 2_end.txt
 mkdir trimmed_reads
 paste -d ' ' 1_end.txt 2_end.txt | xargs -n 2 trim_galore --paired --fastqc -o trimmed_reads
 
-######## --------------> Delete raw reads for space
-rm -rf *fastq.gz
+######## --------------> Delete raw reads for space (if-fi safeguard in case trimming fails)
+if [-f trimmed/*gastq.gz]
+	rm -rf *fastq.gz
+fi
+
 ###################################  ALIGNMENT #####################################################
 ######## --------------> Retrieve hg38
 if [ ! -f "hg38.fa" ]; then
@@ -67,8 +71,10 @@ bowtie2 -x hg38_index/hg38 \
     --threads 8 | samtools view -@8 -bS | samtools sort -@8 -o aligned_reads/aligned.sorted.bam
 
 
-######## --------------> Delete trimmed reads & other files for space
-rm -rf trimmed_reads 1_end.txt 2_end.txt
+######## --------------> Delete trimmed reads & other files for space (if-fi safegaurd if alignment fails)
+if [-f aligned_reads/*]
+	rm -rf trimmed_reads 1_end.txt 2_end.txt
+fi
 
 ######## --------------> Get GTF/GFF
 if [ ! -f "hg38.refGene.gtf" ]; then
